调整公用post 请求 完成对接发送私信给用户 视频两个接口调通

+++ DYSDK/Model/BaseData.cs
@@ -23,7 +23,7 @@ namespace DYSDK.Model
         /// 用于下一页请求的cursor
         /// </summary>
         [JsonProperty("cursor")]
-        public int Cursor { get; set; }
+        public long Cursor { get; set; }
 
         /// <summary>
         /// 是否还有更多数据
